refactor: Add useShowAddressBook hook
