+++ packages/kit/src/views/Setting/pages/List/DefaultSection/index.tsx
@@ -4,15 +4,12 @@ import { useIntl } from 'react-intl';
 
 import { YStack } from '@onekeyhq/components';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { showAddressSafeNotificationDialog } from '@onekeyhq/kit/src/components/AddressInput/AddressSafeDialog';
 import { ListItem } from '@onekeyhq/kit/src/components/ListItem';
 import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
 import { usePromiseResult } from '@onekeyhq/kit/src/hooks/usePromiseResult';
+import { useShowAddressBook } from '@onekeyhq/kit/src/hooks/useShowAddressBook';
 import { useBackupEntryStatus } from '@onekeyhq/kit/src/views/CloudBackup/components/useBackupEntryStatus';
-import {
-  useAddressBookPersistAtom,
-  usePasswordPersistAtom,
-} from '@onekeyhq/kit-bg/src/states/jotai/atoms';
+import { usePasswordPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import {
   EAppEventBusNames,
   appEventBus,
@@ -24,7 +21,6 @@ import {
   ECloudBackupRoutes,
   EDAppConnectionModal,
   ELiteCardRoutes,
-  EModalAddressBookRoutes,
   EModalKeyTagRoutes,
   EModalRoutes,
 } from '@onekeyhq/shared/src/routes';
@@ -47,22 +43,9 @@ export const useOnLock = () => {
 
 const AddressBookItem = () => {
   const intl = useIntl();
-  const navigation = useAppNavigation();
-  const showAddressBook = useCallback(async () => {
-    await backgroundApiProxy.servicePassword.promptPasswordVerify();
-    navigation.push(EModalAddressBookRoutes.ListItemModal);
-    defaultLogger.setting.page.enterAddressBook();
-  }, [navigation]);
-  const [{ hideDialogInfo }] = useAddressBookPersistAtom();
-  const onPress = useCallback(async () => {
-    if (!hideDialogInfo) {
-      await showAddressSafeNotificationDialog({
-        intl,
-      });
-      await backgroundApiProxy.serviceAddressBook.hideDialogInfo();
-    }
-    await showAddressBook();
-  }, [hideDialogInfo, showAddressBook, intl]);
+  const onPress = useShowAddressBook({
+    useNewModal: false,
+  });
   return (
     <ListItem
       icon="ContactsOutline"
